feat(cmd): add exclude option to summary command
I don't want to use regex for this, since glob patterns are user friendlier.
And I'm not using filepath.Match because it only works on file names,
not on the entire path, and the syntax is Go-specific, which is not as
widely familiar as Git's ignore syntax.

filepath.Match also doesn't support the negation feature (e.g.
'*,!*captchav2*' should exclude everything *but* files matching
'*captchav2*'). I need this feature to implement allow lists, i.e.
exclude everything except the files/packages modified in a PR.