feat(cmd): add exclude output option to summary command
This option is different from --exclude, since it doesn't apply to the
coverage calculation. That is, if a package or file is excluded with
--exclude-output, but not with --exclude, it wouldn't affect the actual
coverage percent. This is important, because we want to show only
specific files/packages modified in a PR, but the actual coverage should
be calculated with all other files included.