refactor: add app package and Context, move cli
This moves the CLI implementation to a separate package, as a subpackage
of a new root-level app package. The idea is that the app contains all
dependencies and shared context, using interfaces for external systems.
This keeps the main() function small and trivial, and makes the CLI
testable, since in-memory mocks can be easily passed.

It introduces the mandelsoft/vfs package, which offers convenient
fileystem interfaces. Unlike io/fs from stdlib, it allows writing as
well, which is convenient for testing.