chore: vendor dependencies
This vendors all fcov dependencies.

There are several advantages to vendoring:
- Avoids dependencies on external servers, including Go proxies. This
  makes builds (more) reproducible, and potentially faster, which is
  important in CI.
- Gives visibility into all the code it takes to build the application.
  We don't necessarily have to review it in depth, but it gives more
  insight about the size of a dependency, which might make us reconsider
  depending on it.

The drawback is that it makes the repo larger, which for simple projects
such as this isn't a concern. And that it clutters PRs of actual app
changes, but this shouldn't be an issue if any vendor changes are done
in separate commits. The vendor directory can also be excluded when
reviewing locally with: `git diff -- . ':!vendor'`.